false_prophet/requirements.txt
boto==2.49.0
boto3==1.9.211
botocore==1.12.211
certifi==2019.6.16
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
gensim==3.8.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
nltk==3.4.4
numpy==1.17.0
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.1
six==1.12.0
smart-open==1.8.4
urllib3==1.25.3
Werkzeug==0.15.5
